style(headless): moved ssr files next to their csr counterparts
https://coveord.atlassian.net/browse/KIT-2714
