+++ packages/headless/src/controllers/result-list/headless-result-list.ssr.ts
@@ -1,17 +1,17 @@
-import {SearchEngine} from '../../../app/search-engine/search-engine';
-import {ControllerDefinitionWithoutProps} from '../../../app/ssr-engine/types/common';
+import {SearchEngine} from '../../app/search-engine/search-engine';
+import {ControllerDefinitionWithoutProps} from '../../app/ssr-engine/types/common';
 import {
   ResultList,
   ResultListProps,
   buildResultList,
-} from '../../result-list/headless-result-list';
+} from './headless-result-list';
 
 export type {
   ResultListOptions,
   ResultListProps,
   ResultListState,
   ResultList,
-} from '../../result-list/headless-result-list';
+} from './headless-result-list';
 
 /**
  * @internal
