+++ packages/headless/src/ssr.index.ts
@@ -15,19 +15,23 @@ export type {
   InferCSRState,
 } from './app/ssr-engine/types/core-engine';
 
-export {defineSearchEngine} from './app/ssr-engine/ssr-engine';
+export type {
+  SearchEngineDefinition,
+  SearchEngineDefinitionOptions,
+} from './app/search-engine/search-engine.ssr';
+export {defineSearchEngine} from './app/search-engine/search-engine.ssr';
 
 export type {
   SearchBox,
   SearchBoxOptions,
   SearchBoxProps,
   SearchBoxState,
-} from './controllers/ssr/search-box/headless-ssr-search-box';
-export {defineSearchBox} from './controllers/ssr/search-box/headless-ssr-search-box';
+} from './controllers/search-box/headless-search-box.ssr';
+export {defineSearchBox} from './controllers/search-box/headless-search-box.ssr';
 export type {
   ResultList,
   ResultListOptions,
   ResultListProps,
   ResultListState,
-} from './controllers/ssr/result-list/headless-ssr-result-list';
-export {defineResultList} from './controllers/ssr/result-list/headless-ssr-result-list';
+} from './controllers/result-list/headless-result-list.ssr';
+export {defineResultList} from './controllers/result-list/headless-result-list.ssr';
